+++ src/Attributes/NestedAttributes/NestedDateAttribute.php
@@ -27,7 +27,7 @@ class NestedDateAttribute extends DateAttribute
 
     public function getSearchCondition(Query $query, $table, $value, $searchmode, $fieldname = '')
     {
-        if (!$this->getOwnerInstance()->isNestedAttribute($this->fieldName())) {
+        if (!$this->getOwnerInstance()->hasNestedAttribute($this->fieldName())) {
             return parent::getSearchCondition($query, $table, $value, $searchmode, $fieldname);
         }
         return parent::getSearchCondition($query, $table, $value, $searchmode, $this->buildSQLSearchValue($table));