+++ src/Attributes/NestedAttributes/NestedListAttribute.php
@@ -28,7 +28,7 @@ class NestedListAttribute extends ListAttribute
 
     public function getSearchCondition(Query $query, $table, $value, $searchmode, $fieldname = '')
     {
-        if (!$this->getOwnerInstance()->isNestedAttribute($this->fieldName())) {
+        if (!$this->getOwnerInstance()->hasNestedAttribute($this->fieldName())) {
             return parent::getSearchCondition($query, $table, $value, $searchmode, $fieldname);
         }
         // We only support 'exact' matches.