+++ src/Core/Menu/HeaderItem.php
@@ -6,20 +6,14 @@ namespace Sintattica\Atk\Core\Menu;
 
 class HeaderItem extends Item
 {
-
-    /**
-     * HeaderItem constructor.
-     * @param string $name
-     */
     public function __construct(string $name)
     {
         parent::__construct();
         $this->name = $name;
     }
 
-
     protected function createIdentifierComponents(): ?string
     {
-        return "";
+        return '';
     }
 }